# app/openapi.yml
openapi: 3.0.3
info:
  title: Tom Hanks Catalog
  version: "1.0"
  description: |
    Rotas públicas do catálogo. É uma aplicação web tradicional: as rotas devolvem páginas HTML
    e os formulários são enviados como `application/x-www-form-urlencoded` (ou `multipart/form-data`
    no upload da foto). Depois de uma ação que dá certo, a resposta é um redirecionamento `302`.

    **Pra usar o "Try it out":** faça login no site nesse mesmo navegador (ou pelo próprio
    `POST /login` aqui embaixo). A sessão fica no cookie `session`, que o navegador já manda junto.
    Sem login, as rotas protegidas redirecionam pra `/login` — como o navegador segue o
    redirecionamento sozinho, o Swagger mostra o HTML da página de login.

    O auth-service e o log-service têm documentação própria, em `/apidocs` de cada um
    (só acessível rodando localmente, já que eles não têm porta pública).
...
components:
  securitySchemes:
    sessao:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sessão criado no login.
  responses:
    Html:
      description: Página HTML.
      content:
        text/html:
          schema:
            type: string
...
    CampoFaltando:
      description: Faltou um campo obrigatório do formulário.
...
paths:
  /cadastro:
    get:
      tags: [contas]
      summary: Formulário de cadastro
      responses:
        "200":
          $ref: "#/components/responses/Html"
    post:
      tags: [contas]
      summary: Cria uma conta
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [nome, email, senha]
              properties:
                nome:
                  type: string
                  example: Maria
                email:
                  type: string
                  example: maria@exemplo.com
                senha:
                  type: string
                  format: password
      responses:
        "302":
          description: Conta criada. Redireciona pra `/login`.
          headers:
            Location:
              schema:
                type: string
                example: /login
        "200":
          description: >
            Não cadastrou (campo vazio, e-mail já usado ou auth-service fora do ar).
            Volta o formulário com a mensagem de erro.
          content:
            text/html:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/CampoFaltando"
  /login:
    get:
      tags: [contas]
      summary: Formulário de login
      responses:
        "200":
          $ref: "#/components/responses/Html"
    post:
      tags: [contas]
      summary: Faz login
      description: Confere a senha no auth-service e, se estiver certa, cria o cookie de sessão.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [email, senha]
              properties:
                email:
                  type: string
                  example: maria@exemplo.com
                senha:
                  type: string
                  format: password
      responses:
        "302":
          description: Login certo. Cria o cookie `session` e redireciona pro catálogo.
          headers:
            Location:
              schema:
                type: string
                example: /
            Set-Cookie:
              schema:
                type: string
                example: session=eyJub21lIjoi...; HttpOnly; Path=/
        "200":
          description: E-mail ou senha errados. Volta o formulário com a mensagem de erro.
          content:
            text/html:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/CampoFaltando"
...
  /esqueci-senha:
    get:
      tags: [contas]
      summary: Formulário de esqueci a senha
      responses:
        "200":
          $ref: "#/components/responses/Html"
    post:
      tags: [contas]
      summary: Pede o link de redefinição por e-mail
      description: >
        A resposta é sempre a mesma, com ou sem conta cadastrada nesse e-mail, pra não dar pra
        descobrir quem tem conta.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  example: maria@exemplo.com
      responses:
        "200":
          description: Página com a mensagem "Se esse e-mail estiver cadastrado, você vai receber um link".
          content:
            text/html:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/CampoFaltando"
  /resetar-senha:
    get:
      tags: [contas]
      summary: Formulário de nova senha (link do e-mail)
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Html"
    post:
      tags: [contas]
      summary: Troca a senha
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token, nova_senha]
              properties:
                token:
                  type: string
                nova_senha:
                  type: string
                  format: password
      responses:
        "302":
          description: Senha trocada. Redireciona pra `/login?redefinida=1`.
        "200":
          description: Token inválido, já usado ou expirado. Volta o formulário com a mensagem de erro.
          content:
            text/html:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/CampoFaltando"
...